requirements.txt
click==6.6
Flask==0.11.1
Flask-Cache==0.13.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
requests==2.10.0
waitress==0.9.0
Werkzeug==0.11.10
